fix njupt posts with url redirections (#8961)
* fix njupt posts with url redirections

skip posts that are redirected to error pages due to access control

* fix: migrate to v2

* fix: pubDate depreciated url.resolve
