+++ lib/router.js
@@ -607,9 +607,6 @@ router.get('/ustc/job/:category?', lazyloadRouteHandler('./routes/universities/u
 // UTdallas ISSO
 router.get('/utdallas/isso', lazyloadRouteHandler('./routes/universities/utdallas/isso'));
 
-// 南京邮电大学
-router.get('/njupt/jwc/:type?', lazyloadRouteHandler('./routes/universities/njupt/jwc'));
-
 // 南昌航空大学
 router.get('/nchu/jwc/:type?', lazyloadRouteHandler('./routes/universities/nchu/jwc'));
