+++ lib/v2/njupt/maintainer.js
@@ -1,0 +1,3 @@
+module.exports = {
+    '/jwc/:type?': ['shaoye'],
+};
