+++ lib/v2/njupt/router.js
@@ -1,0 +1,3 @@
+module.exports = (router) => {
+    router.get('/jwc/:type?', require('./jwc'));
+};
